Implementando segurança

+++ src/temas/services/tema.service.ts
@@ -3,7 +3,6 @@ import { InjectRepository } from "@nestjs/typeorm";
 import { DeleteResult, ILike, Repository } from "typeorm";
 import { Tema } from "../entities/tema.entity";
 
-//tema.service tem os a lógica dos métodos e a repository para executar
 @Injectable()
 export class TemaService {
     constructor(
@@ -13,8 +12,8 @@ export class TemaService {
 
     async findAll(): Promise<Tema[]> {
         return await this.temaRepository.find({
-            relations: {        //relations tem como função exibir os Objetos da Classe Postagem que estão relacionados com os Objetos da Classe Temas
-                postagem: true             //isso acontece com os métodos findAll, findById e findbyDescricao e quando um Objeto é persistido
+            relations: {
+                postagem: true
             }
         });
     }
@@ -53,14 +52,20 @@ export class TemaService {
 
     async update(tema: Tema): Promise<Tema> {
 
-        await this.findById(tema.id);
+        let buscaTema = await this.findById(tema.id);
+
+        if (!buscaTema || !tema.id)
+            throw new HttpException('Tema não encontrado!', HttpStatus.NOT_FOUND);
 
         return await this.temaRepository.save(tema);
     }
 
     async delete(id: number): Promise<DeleteResult> {
 
-        await this.findById(id);
+        let buscaTema = await this.findById(id);
+
+        if (!buscaTema)
+            throw new HttpException('Tema não encontrado!', HttpStatus.NOT_FOUND);
 
         return await this.temaRepository.delete(id);
